+++ src/38query.js
@@ -22,8 +22,14 @@ function queryfn(query,oldscope,cb, A,B) {
 //			var res = flatArray(q(query.params,null,queryfn2,(-idx-1),query));
 
 //			var res = flatArray(queryfn(q.query,null,queryfn2,(-idx-1),query));
-
+//			console.log(A,B);
+// console.log(q);
+			q.query.params = query.params;
+//			query.queriesdata[idx] = 
 			queryfn(q.query,null,queryfn2,(-idx-1),query);
+
+
+//			console.log(27,q);
 
 
 //			query.explaination.push({explid: query.explid++, description:'Query '+idx,ms:Date.now()-ms});
@@ -89,6 +95,7 @@ function queryfn2(data,idx,query) {
 	} else {
 		// subqueries
 		query.queriesdata[-idx-1] = flatArray(data);
+		console.log(98,query.queriesdata);
 //		console.log(79,query.queriesdata);
 	}